repo_data_files/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0

transformers>=4.30.0
albumentations>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.62.0

nltk>=3.6.0
spacy>=3.1.0
scipy>=1.7.0
networkx>=2.6.0

plotly>=5.3.0
wandb>=0.12.0
tensorboard>=2.8.0

fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0
grpcio>=1.41.0
protobuf>=3.19.0

sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0
pymongo>=4.0.0

prometheus-client>=0.12.0
structlog>=21.5.0
python-json-logger>=2.0.0

pytest>=6.2.0
pytest-cov>=2.12.0
pytest-asyncio>=0.15.0
pytest-mock>=3.6.0

black>=21.7b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910
pre-commit>=2.15.0

sphinx>=4.1.0
sphinx-rtd-theme>=0.5.0
mkdocs>=1.2.0
mkdocs-material>=8.0.0

docker>=6.1.0
kubernetes>=26.1.0
helm>=3.7.0
kubectl>=1.22.0

cryptography>=3.4.0
python-jose>=3.3.0
passlib>=1.7.0
python-multipart>=0.0.5

numba>=0.54.0
cupy-cuda11x>=11.0.0
dask>=2021.9.0
ray>=1.9.0

python-dotenv>=0.19.0
pyyaml>=5.4.0
tqdm>=4.62.0
colorama>=0.4.4
rich>=10.12.0 